Removal of libc directory and added more functions

+++ ft_strncat.c
@@ -6,7 +6,7 @@
 /*   By: lyoung <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/02/28 14:44:17 by lyoung            #+#    #+#             */
-/*   Updated: 2017/02/28 14:57:53 by lyoung           ###   ########.fr       */
+/*   Updated: 2017/03/01 13:40:33 by lyoung           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strncat(char *s1, const char *s2, size_t n)
 {
-	int		i;
-	int		j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	j = 0;
